requirements.txt
appdirs==1.4.3
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
distlib==0.3.0
Django==3.0.4
django-filter==2.2.0
djangorestframework==3.11.0
filelock==3.0.12
idna==2.9
numpy==1.18.2
panda==0.3.1
pandas==1.0.3
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
virtualenv==20.0.14
